backend/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.23.0
numpy==1.20.2
opencv-python==4.5.1.48
pymongo==3.11.3
python-dotenv==0.16.0
six==1.15.0
Werkzeug==1.0.1
